user roles and notification improved

+++ resources/views/questions/show.blade.php
@@ -8,7 +8,6 @@
 
             <hr>
             <h3>{{ count($solutions) }} Solutions</h3>
-
             @foreach($solutions as $solution)
                 <div class="row" id="vote">
                     <div class="col-md-2">
@@ -16,14 +15,7 @@
                             {!! Form::model($solution, ['route' => ['vote.increase', $solution->id], 'method' => 'PUT']) !!}
                             {{ Form::submit('&uarr;', array('class' => 'btn btn-default')) }}
                             {!! Form::close() !!}
-                            {{--<button type="button" class="btn btn-default">--}}
-                                {{--<span class="glyphicon glyphicon-triangle-top" aria-hidden="true"></span>--}}
-                            {{--</button>--}}
                             <h4>{{ $solution->votes }}</h4>
-                            {{--<button type="button" class="btn btn-default">--}}
-                                {{--<span class="glyphicon glyphicon-triangle-bottom" aria-hidden="true"></span>--}}
-                            {{--</button>--}}
-
                             {!! Form::model($solution, ['route' => ['vote.decrease', $solution->id], 'method' => 'PUT']) !!}
                             {{ Form::submit('&darr;', array('class' => 'btn btn-default')) }}
                             {!! Form::close() !!}
@@ -40,7 +32,7 @@
                         <div class="well" id="solution-info">
                             <div style="display: flex;">
                                 @if(Auth::check())
-                                    @if(Auth::user()->id == $solution->user_id || strcmp(Auth::user()->type, 'admin') == 0)
+                                    @if(Auth::user()->id == $solution->user_id || strcmp(Auth::user()->role, 'admin') == 0)
                                         {!! Html::linkRoute('solutions.edit', 'Edit', array($solution->id), array('class'=> 'btn btn-primary', 'style' => 'margin: 3px;')) !!}
                                         {!! Form::model($solution, ['route' => ['solutions.destroy', $solution->id], 'method' => 'DELETE']) !!}
                                         {{ Form::submit('Delete', array('class' => 'btn btn-danger', 'style' => 'margin: 3px;')) }}
@@ -58,6 +50,7 @@
             @endforeach
 
             @if(Auth::check())
+
                 {!! Form::open(['route' => 'solutions.store']) !!}
                 {{ Form::label('body', 'Body:') }}
                 {{ Form::textarea('body', null, array('class' => 'form-control')) }}
@@ -74,27 +67,29 @@
             <div class="col-md-3 col-md-offset-1">
                 <div class="well">
                     <dl class="dl-horizontal">
-                        <label>Question Created At :</label>
-                        <p>{{$question->created_at}}</p>
+                        <label>Question Asked</label>
+                        <p>{{\App\Http\Controllers\QuestionController::time_elapsed_string($question->created_at)}}</p>
                     </dl>
                     <dl class="dl-horizontal">
                     <label>Views :</label>
                     <p>{{$question->view}}</p>
                     </dl>
-                    {{--<dl class="dl-horizontal">--}}
-                        {{--<label>Post URL :</label>--}}
-                        {{--<p>{{url($question->slug)}}</p>--}}
-                    {{--</dl>--}}
-                    {{--<hr>--}}
                     <div class="row">
-                        <div class="col-md-6">
+
+
+{{--                        {{dd(Auth::user()->)}}--}}
+                    @if(Auth::check())
+                        @if(Auth::user()->id == $question->user_id || strcmp(Auth::user()->role, 'admin') == 0)
+                            <div class="col-md-6">
                             {!! Html::linkRoute('questions.edit', 'Edit', array($question->id), array('class'=> 'btn btn-primary btn-lg btn-block')) !!}
-                        </div>
-                        <div class="col-md-6">
-                            {!! Form::model($question, ['route' => ['questions.destroy', $question->id], 'method' => 'DELETE']) !!}
-                            {{ Form::submit('Delete', array('class' => 'btn btn-danger btn-lg btn-block')) }}
-                            {!! Form::close() !!}
-                        </div>
+                            </div>
+                            <div class="col-md-6">
+                                {!! Form::model($question, ['route' => ['questions.destroy', $question->id], 'method' => 'DELETE']) !!}
+                                {{ Form::submit('Delete', array('class' => 'btn btn-danger btn-lg btn-block')) }}
+                                {!! Form::close() !!}
+                            </div>
+                        @endif
+                    @endif
                     </div>
                 </div>
             </div>
@@ -102,14 +97,7 @@
     </div>
 
     <script>
-
-        {{--function increase(solution_id){--}}
-            {{--{{$id = solution_id}}--}}
-            {{--window.location = {{ route('vote.increase', $id) }};--}}
-        {{--}--}}
-
         var clientHeight = document.getElementById('vote').clientHeight;
-//        console.log(clientHeight);
         document.getElementById('solution-body').style.minHeight =clientHeight-20 + "px";
         document.getElementById('solution-info').style.minHeight =clientHeight-20 + "px";
     </script>
